:recycle: [#2661] Minor cleanups after rebase on master
* move service fetch modal to ActionFetchFromService component
* fix displayName for FormLogicContext
* add FormLogicContext in FormLogic component instead of FormCreationForm
